Use hiera for more things

+++ modules/ocf/manifests/packages/postfix.pp
@@ -5,35 +5,32 @@ class ocf::packages::postfix {
     before => Package['postfix'],
   }
 
-  package {
-    'postfix':
-    ;
-    'bsd-mailx':
-      require => Package['postfix'],
-    ;
-  }
+  if !tagged('ocf_mail') {
+    package {
+      ['postfix', 'bsd-mailx']:;
+    }
 
-  # disable incoming TCP/IP connections
-  augeas { '/etc/postfix/master.cf':
-    context => '/files/etc/postfix/master.cf',
-    changes => 'rm *[type=\'inet\']',
-    notify  => Service['postfix'],
-    require => Package['postfix'],
-  }
-
-  # mail configuration
-  file {
-    '/etc/mailname':
-      content => 'ocf.berkeley.edu',
-    ;
-    '/etc/postfix/main.cf':
-      source  => 'puppet:///modules/ocf/postfix/main.cf',
+    # disable incoming TCP/IP connections
+    augeas { '/etc/postfix/master.cf':
+      context => '/files/etc/postfix/master.cf',
+      changes => 'rm *[type=\'inet\']',
       notify  => Service['postfix'],
       require => Package['postfix'],
-    ;
-  }
+    }
 
-  service { 'postfix':
-    require => Package['postfix'],
+    # mail configuration
+    file {
+      '/etc/mailname':
+        content => "ocf.berkeley.edu\n";
+
+      '/etc/postfix/main.cf':
+        source  => 'puppet:///modules/ocf/postfix/main.cf',
+        notify  => Service['postfix'],
+        require => Package['postfix'];
+    }
+
+    service { 'postfix':
+      require => Package['postfix'],
+    }
   }
 }
